+++ core/rate_limit.py
@@ -1,0 +1,23 @@
+import time
+import asyncio
+import logging
+from functools import wraps
+
+logger = logging.getLogger("agent-runner")
+
+def rate_limit(calls_per_minute=20):
+    min_interval = 60.0 / calls_per_minute
+    last_call_time = 0
+
+    def decorator(func):
+        @wraps(func)
+        async def wrapper(*args, **kwargs):
+            nonlocal last_call_time
+            now = time.time()
+            delta = now - last_call_time
+            if delta < min_interval:
+                await asyncio.sleep(min_interval - delta)
+            last_call_time = time.time()
+            return await func(*args, **kwargs)
+        return wrapper
+    return decorator
